Allow normalising redirects
